Optimized the implementation of NFmiPath::PacificView
The code now inspects the bounding box of subpaths to see whether a simple
operation is sufficient to translate the subpath to a Pacific view or whether
it must be cut into pieces and rebuilt. The previous implementation took about
two seconds to process ESRI World data, the new one takes about 0.3 seconds.
The time is still significant since the entire Eurasia has to be cut into pieces.